-- Successfully implemented ItemTouchHelper with the swipe gesture when user swipes to the right the transaction is removed from the list and from the DB(small issue if index is 5+) -- Updated the TransactionId type to long to be able to work with the ItemTouchHelper -- polished the view a little bit (adding drawables, nestedScrollView, theming and styling)

+++ app/src/main/java/com/zanoapp/applediseaseIdentification/localDataPersistence/transactionsDB/Transaction.kt
@@ -11,7 +11,7 @@ import java.sql.Date
 @Entity
 data class Transaction (
     @PrimaryKey(autoGenerate = true)
-        public val transactionId: Int,
+        public val transactionId: Long,
 
     @ColumnInfo(name = "transaction_type")
         public val transactionType: String,
